refactor(validation): consolidate python-version-consistency into canonical module
Migrates all logic from scripts_lib/check_python_version_consistency.py
into hephaestus/validation/python_version.py (the already-registered
console-script entry point). Adds section-bounded regex for mypy
extraction (fixes latent re.DOTALL cross-section bug), 8 new public
functions, CI checks integrated into main(), and a delegation shim in
scripts_lib for backward compatibility.

Closes #1189
