+++ hephaestus/validation/__init__.py
@@ -45,7 +45,18 @@ from hephaestus.validation.markdown import (
     validate_relative_link,
 )
 from hephaestus.validation.mypy_per_file import check_mypy_per_file
-from hephaestus.validation.python_version import check_python_version_consistency
+from hephaestus.validation.python_version import (
+    check_ci_matrix_coverage,
+    check_pixi_python_ceiling,
+    check_project_version_consistency,
+    check_python_version_consistency,
+    extract_ci_matrix_python_versions,
+    extract_classifiers_python_versions,
+    extract_pixi_python_ceiling,
+    extract_pixi_workspace_version,
+    extract_project_version,
+    extract_pyproject_versions_str,
+)
 from hephaestus.validation.readme_commands import (
     CodeBlock,
     ReadmeValidator,
@@ -75,12 +86,15 @@ __all__ = [
     "TierLabelFinding",
     "ValidationReport",
     "ValidationResult",
+    "check_ci_matrix_coverage",
     "check_coverage",
     "check_doc_config_consistency",
     "check_markdown_formatting",
     "check_max_complexity",
     "check_mypy_per_file",
     "check_no_loose_test_files",
+    "check_pixi_python_ceiling",
+    "check_project_version_consistency",
     "check_python_version_consistency",
     "check_required_sections",
     "check_stale_scripts",
@@ -88,7 +102,13 @@ __all__ = [
     "check_test_structure",
     "count_markdown_issues",
     "detect_shadowing",
+    "extract_ci_matrix_python_versions",
+    "extract_classifiers_python_versions",
     "extract_markdown_links",
+    "extract_pixi_python_ceiling",
+    "extract_pixi_workspace_version",
+    "extract_project_version",
+    "extract_pyproject_versions_str",
     "extract_sections",
     "filter_audit_results",
     "find_readmes",
