+++ hephaestus/validation/python_version.py
@@ -26,6 +26,8 @@ tomllib = import_tomllib()
 
 _CLASSIFIER_VERSION_RE = re.compile(r"Programming Language :: Python :: (\d+\.\d+)$")
 _DOCKERFILE_FROM_RE = re.compile(r"^\s*FROM\s+python:(\d+\.\d+)", re.IGNORECASE | re.MULTILINE)
+_CI_MATRIX_PYTHON_RE = re.compile(r"python-version:\s*\[([^\]]+)\]")
+_PIXI_PYTHON_BOUND_RE = re.compile(r"<=?\s*(\d+\.\d+)")
 
 
 def extract_pyproject_versions(pyproject_path: Path) -> dict[str, str]:
@@ -102,6 +104,22 @@ def _extract_via_tomllib(pyproject_path: Path) -> dict[str, str]:
 def _extract_via_regex(pyproject_path: Path) -> dict[str, str]:
     """Fallback: extract versions using regex (no TOML parser available)."""
     content = pyproject_path.read_text(encoding="utf-8")
+    return _extract_versions_from_text(content)
+
+
+def _extract_versions_from_text(content: str) -> dict[str, str]:
+    """Core regex extraction from raw pyproject.toml text.
+
+    Uses section-bounded negative-lookahead for mypy to avoid crossing into
+    adjacent [tool.*] sections (same contract as scripts_lib original).
+
+    Args:
+        content: Raw text content of a pyproject.toml file.
+
+    Returns:
+        Dict mapping source labels to version strings.
+
+    """
     versions: dict[str, str] = {}
 
     match = re.search(r'requires-python\s*=\s*"([^"]+)"', content)
@@ -110,7 +128,11 @@ def _extract_via_regex(pyproject_path: Path) -> dict[str, str]:
         if ver_match:
             versions["requires-python"] = ver_match.group(1)
 
-    match = re.search(r'\[tool\.mypy\].*?python_version\s*=\s*"([^"]+)"', content, re.DOTALL)
+    # Section-bounded: does NOT cross into the next [tool.*] section.
+    match = re.search(
+        r'\[tool\.mypy\]\n(?:(?!\[).+\n)*?python_version\s*=\s*"([^"]+)"',
+        content,
+    )
     if match:
         versions["mypy.python_version"] = match.group(1)
 
@@ -122,6 +144,24 @@ def _extract_via_regex(pyproject_path: Path) -> dict[str, str]:
         versions["ruff.target-version"] = f"{match.group(1)}.{match.group(2)}"
 
     return versions
+
+
+def extract_pyproject_versions_str(content: str) -> dict[str, str]:
+    """Extract Python version specs from raw pyproject.toml text (string API).
+
+    Equivalent to :func:`extract_pyproject_versions` but accepts the file
+    content as a string rather than a :class:`~pathlib.Path`. Used by the
+    ``scripts_lib`` shim and callers that already have the file content in
+    memory.
+
+    Args:
+        content: Raw text content of a pyproject.toml file.
+
+    Returns:
+        Dict mapping source labels to version strings.
+
+    """
+    return _extract_versions_from_text(content)
 
 
 def get_dockerfile_python_version(dockerfile_path: Path) -> str | None:
@@ -141,6 +181,247 @@ def get_dockerfile_python_version(dockerfile_path: Path) -> str | None:
     content = dockerfile_path.read_text(encoding="utf-8")
     m = _DOCKERFILE_FROM_RE.search(content)
     return m.group(1) if m else None
+
+
+def extract_project_version(content: str) -> str | None:
+    """Extract version from the [project] section in pyproject.toml content.
+
+    Uses a section-bounded regex with negative lookahead to prevent matching
+    version keys that belong to a different TOML section.
+
+    Args:
+        content: The raw text content of a pyproject.toml file.
+
+    Returns:
+        The version string if found within [project], or None.
+
+    """
+    match = re.search(
+        r'\[project\]\n(?:(?!\[).+\n)*?version\s*=\s*"([^"]+)"',
+        content,
+    )
+    return match.group(1) if match else None
+
+
+def extract_pixi_workspace_version(content: str) -> str | None:
+    """Extract version from the [workspace] section in pixi.toml content.
+
+    Uses a section-bounded regex with negative lookahead to prevent matching
+    version keys that belong to a different TOML section.
+
+    Args:
+        content: The raw text content of a pixi.toml file.
+
+    Returns:
+        The version string if found within [workspace], or None.
+
+    """
+    match = re.search(
+        r'\[workspace\]\n(?:(?!\[).+\n)*?version\s*=\s*"([^"]+)"',
+        content,
+    )
+    return match.group(1) if match else None
+
+
+def extract_pixi_python_ceiling(content: str) -> str | None:
+    """Extract the upper-bound minor from pixi.toml ``[dependencies] python``.
+
+    Parses e.g. ``python = ">=3.10,<3.14"`` and returns the exclusive (``<``)
+    or inclusive (``<=``) upper bound ``"3.14"``. Returns ``None`` if there is
+    no ``python`` key in the base ``[dependencies]`` table, or it carries no
+    ``<``/``<=`` bound.
+
+    The regex is bounded to the base ``[dependencies]`` table (it does not
+    cross into the next ``[`` section header), so a ``python`` pin in a
+    ``[feature.*.dependencies]`` table is intentionally not matched.
+
+    Args:
+        content: The raw text content of a pixi.toml file.
+
+    Returns:
+        The upper-bound ``"major.minor"`` string, or ``None``.
+
+    """
+    match = re.search(
+        r'\[dependencies\]\n(?:(?!\[).+\n)*?python\s*=\s*"([^"]+)"',
+        content,
+    )
+    if not match:
+        return None
+    bound = _PIXI_PYTHON_BOUND_RE.search(match.group(1))
+    return bound.group(1) if bound else None
+
+
+def extract_classifiers_python_versions(content: str) -> list[str]:
+    """Extract Python X.Y version strings from pyproject.toml classifiers.
+
+    Looks for entries of the form:
+        "Programming Language :: Python :: 3.10"
+
+    Args:
+        content: The raw text content of a pyproject.toml file.
+
+    Returns:
+        Sorted list of version strings (e.g. ["3.10", "3.11"]).
+
+    """
+    versions = re.findall(
+        r'"Programming Language :: Python :: (\d+\.\d+)"',
+        content,
+    )
+    return sorted(set(versions))
+
+
+def extract_ci_matrix_python_versions(content: str) -> list[str]:
+    """Extract Python version strings from a GitHub Actions workflow file.
+
+    Looks for a ``python-version`` matrix key with an inline list, e.g.::
+
+        python-version: ["3.10", "3.11", "3.12"]
+
+    Args:
+        content: The raw YAML text of a GitHub Actions workflow file.
+
+    Returns:
+        Sorted list of version strings (e.g. ["3.10", "3.11"]), or empty list
+        if no ``python-version`` matrix key is found.
+
+    """
+    match = _CI_MATRIX_PYTHON_RE.search(content)
+    if not match:
+        return []
+    versions = re.findall(r'["\']?(\d+\.\d+)["\']?', match.group(1))
+    return sorted(set(versions))
+
+
+def check_project_version_consistency(repo_root: Path) -> bool:
+    """Check that pyproject.toml and pixi.toml project versions agree.
+
+    Reads the [project].version from pyproject.toml and the [workspace].version
+    from pixi.toml (if it exists) and verifies they match. If pixi.toml has no
+    version field the check passes — that is the expected state for this project.
+
+    Args:
+        repo_root: Repository root directory.
+
+    Returns:
+        True if versions are consistent (or pixi.toml has no version), False otherwise.
+
+    """
+    pyproject_path = repo_root / "pyproject.toml"
+    if not pyproject_path.exists():
+        return True
+    project_version = extract_project_version(pyproject_path.read_text())
+    pixi_path = repo_root / "pixi.toml"
+    if not pixi_path.exists():
+        if project_version:
+            print(f"OK: pyproject.toml project version = {project_version!r} (no pixi.toml)")
+        return True
+    pixi_version = extract_pixi_workspace_version(pixi_path.read_text())
+    if pixi_version is None:
+        if project_version:
+            print(f"OK: pyproject.toml project version = {project_version!r}")
+        print("OK: pixi.toml has no [workspace].version (as expected)")
+        return True
+    if project_version == pixi_version:
+        print(f"OK: project version is consistent at {project_version!r}")
+        return True
+    print("ERROR: project version mismatch between pyproject.toml and pixi.toml!")
+    print(f"  pyproject.toml [project].version = {project_version!r}")
+    print(f"  pixi.toml [workspace].version    = {pixi_version!r}")
+    print("  pyproject.toml is the single source of truth — update pixi.toml to match.")
+    return False
+
+
+def check_ci_matrix_coverage(repo_root: Path) -> bool:
+    """Check that the CI python-version matrix covers all classifier versions.
+
+    Parses pyproject.toml classifiers and .github/workflows/test.yml matrix
+    and reports any classifier versions absent from the CI matrix.
+
+    Args:
+        repo_root: Repository root directory.
+
+    Returns:
+        True if CI matrix covers all classifier versions (or no data to compare),
+        False if any classifier version is missing from the CI matrix.
+
+    """
+    pyproject_path = repo_root / "pyproject.toml"
+    if not pyproject_path.exists():
+        return True
+    classifier_versions = extract_classifiers_python_versions(pyproject_path.read_text())
+    if not classifier_versions:
+        return True
+    ci_workflow = repo_root / ".github" / "workflows" / "test.yml"
+    if not ci_workflow.exists():
+        print(f"INFO: CI workflow not found at {ci_workflow} — skipping matrix check")
+        return True
+    matrix_versions = extract_ci_matrix_python_versions(ci_workflow.read_text())
+    if not matrix_versions:
+        print(f"INFO: No python-version matrix found in {ci_workflow} — skipping matrix check")
+        return True
+    missing = sorted(set(classifier_versions) - set(matrix_versions))
+    extra = sorted(set(matrix_versions) - set(classifier_versions))
+    if not missing:
+        print(f"OK: CI matrix covers all classifier Python versions: {matrix_versions}")
+        return True
+    print("ERROR: CI matrix is missing Python versions listed in pyproject.toml classifiers!")
+    print(f"  Classifiers: {classifier_versions}")
+    print(f"  CI matrix:   {matrix_versions}")
+    print(f"  Missing from CI matrix: {missing}")
+    if extra:
+        print(f"  In CI matrix but not in classifiers: {extra}")
+    return False
+
+
+def check_pixi_python_ceiling(repo_root: Path) -> bool:
+    """Ensure pixi.toml's python upper bound stays within the support matrix.
+
+    The locked dev/lint envs must not resolve above the highest classifier
+    Python version (the support ceiling). An unbounded or too-high pin lets
+    pixi resolve to an untested interpreter — e.g. Python 3.14 free-threaded
+    (cp314t), which is neither in the declared support matrix nor CI-tested
+    (see issue #1184). The accepted upper bound is at most one minor above the
+    highest classifier version (``<3.14`` for a 3.13 ceiling).
+
+    Args:
+        repo_root: Repository root directory.
+
+    Returns:
+        True if the pixi python ceiling is present and within one minor of the
+        highest classifier version (or there is nothing to compare), False if
+        the pin is unbounded or its ceiling exceeds the allowed maximum.
+
+    """
+    from packaging.version import Version
+
+    pyproject_path = repo_root / "pyproject.toml"
+    pixi_path = repo_root / "pixi.toml"
+    if not pyproject_path.exists() or not pixi_path.exists():
+        return True
+    classifiers = extract_classifiers_python_versions(pyproject_path.read_text())
+    if not classifiers:
+        return True
+    highest_supported = max(Version(v) for v in classifiers)
+    max_allowed = Version(f"{highest_supported.major}.{highest_supported.minor + 1}")
+    ceiling = extract_pixi_python_ceiling(pixi_path.read_text())
+    if ceiling is None:
+        print(
+            "ERROR: pixi.toml [dependencies] python has no upper bound — "
+            "the env may resolve to an untested interpreter (see #1184).\n"
+            f'  Add an upper bound, e.g. python = ">={classifiers[0]},<{max_allowed}".'
+        )
+        return False
+    if Version(ceiling) > max_allowed:
+        print(
+            "ERROR: pixi.toml python upper bound is too high!\n"
+            f"  pixi cap: <{ceiling}; highest classifier: {highest_supported}; "
+            f"max allowed cap: <{max_allowed} (one minor above support ceiling)."
+        )
+        return False
+    print(f"OK: pixi.toml python ceiling <{ceiling} is within support matrix (<= {max_allowed})")
+    return True
 
 
 def check_python_version_consistency(
@@ -238,20 +519,34 @@ def main() -> int:
         verbose=args.verbose and not args.json,
     )
 
+    project_version_ok = check_project_version_consistency(repo_root)
+    ci_matrix_ok = check_ci_matrix_coverage(repo_root)
+    pixi_ceiling_ok = check_pixi_python_ceiling(repo_root)
+
+    all_ok = (
+        (consistent or not versions)
+        and project_version_ok
+        and ci_matrix_ok
+        and pixi_ceiling_ok
+    )
+
     if args.json:
         report = {
             "consistent": consistent,
             "versions": versions,
-            "passed": consistent or not versions,
+            "ci_checks": {
+                "project_version_ok": project_version_ok,
+                "ci_matrix_ok": ci_matrix_ok,
+                "pixi_ceiling_ok": pixi_ceiling_ok,
+            },
+            "passed": all_ok,
         }
         print(format_output(report, "json"))
-        if not versions:
-            return 0
-        return 0 if consistent else 1
+        return 0 if all_ok else 1
 
     if not versions:
         print("WARNING: No Python version specs found in pyproject.toml")
-        return 0
+        return 0 if all_ok else 1
 
     if consistent:
         unique = set(versions.values())
@@ -262,16 +557,16 @@ def main() -> int:
         if args.verbose:
             for key, val in sorted(versions.items()):
                 print(f"  {key}: {val}")
-        return 0
+    else:
+        print("ERROR: Python version inconsistency detected!", file=sys.stderr)
+        for key, val in sorted(versions.items()):
+            print(f"  {key}: {val}", file=sys.stderr)
+        print(
+            "\nAll version specifications must agree on the same Python version.",
+            file=sys.stderr,
+        )
 
-    print("ERROR: Python version inconsistency detected!", file=sys.stderr)
-    for key, val in sorted(versions.items()):
-        print(f"  {key}: {val}", file=sys.stderr)
-    print(
-        "\nAll version specifications must agree on the same Python version.",
-        file=sys.stderr,
-    )
-    return 1
+    return 0 if all_ok else 1
 
 
 if __name__ == "__main__":
